pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*", "domains*", "schemas*"]

[project]
name = "the-watchman"
version = "0.1.0"
description = "A distributed knowledge graph system for comprehensive computing environment awareness"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "The Watchman Team"}
]
keywords = ["knowledge-graph", "neo4j", "monitoring", "automation", "ocr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "fastapi~=0.111.0",
    "uvicorn[standard]~=0.29.0",
    "pydantic~=2.7.0",
    "pydantic-settings~=2.2.0",
    "python-multipart~=0.0.9",
    "neo4j~=5.20.0",
    "openai~=1.25.0",
    "httpx~=0.27.0",
    "requests~=2.31.0",
    "pytesseract~=0.3.10",
    "Pillow~=10.3.0",
    "opencv-python~=4.9.0",
    "numpy~=1.26.4",
    "imagehash~=4.3.1",
    "pyautogui~=0.9.54",
    "pynput~=1.7.6",
    "python-xlib~=0.33",
    "mss~=9.0.1",
    "watchdog~=4.0.0",
    "docker>=7.0.0",
    "PyYAML~=6.0.1",
    "python-dotenv~=1.0.1",
    "python-dateutil~=2.9.0",
    "redis~=5.0.4",
    "loguru~=0.7.2",
    "typer~=0.12.3",
    "rich~=13.7.1",
]

[project.urls]
Homepage = "https://github.com/Coldaine/the-watchman"
Repository = "https://github.com/Coldaine/the-watchman"
Documentation = "https://github.com/Coldaine/the-watchman/blob/main/README.md"

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "testcontainers[neo4j]>=4.4.0",
    "black>=24.4.2",
    "ruff>=0.4.2",
    "isort>=5.13.2",
    "mypy>=1.9.0",
    "bandit[toml]>=1.7.8",
    "safety>=3.1.0",
    "pre-commit>=3.7.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.12",
    "types-python-dateutil>=2.9.0",
]

# ===== Black Configuration =====
[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

# ===== Ruff Configuration =====
[tool.ruff]
line-length = 100
target-version = "py311"

# Enable specific rule sets
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
    "PL",   # pylint
]

# Ignore specific rules
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function call in argument defaults
    "PLR0913", # too many arguments
    "PLR2004", # magic value comparison
    "PLR0911", # too many return statements
    "ARG001", # unused function argument (common in FastAPI handlers)
    "ARG002", # unused method argument (common in async methods)
    "F841",   # local variable assigned but never used
    "SIM102", # use single if instead of nested (subjective)
    "SIM103", # return condition directly (subjective)
    "SIM108", # use ternary operator (subjective)
    "SIM110", # use any() instead of for loop (subjective)
    "PLW0603", # using global statement (singleton pattern)
    "PLW1510", # subprocess.run without explicit check (handled case-by-case)
    "PLW2901", # for loop variable overwritten (intentional in some cases)
    "PLC0415", # import not at top level (needed for conditional imports)
    "UP028",  # replace yield with yield from (subjective)
    "UP036",  # version block outdated (needed for backwards compatibility)
]

# Exclude directories
exclude = [
    ".git",
    ".venv",
    "build",
    "dist",
    "__pycache__",
]

[tool.ruff.per-file-ignores]
# Allow unused imports in __init__.py files
"__init__.py" = ["F401"]
# Allow print statements in scripts
"scripts/*" = ["T201"]

[tool.ruff.mccabe]
max-complexity = 10

# ===== isort Configuration =====
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true

# ===== mypy Configuration =====
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start permissive, tighten later
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
ignore_missing_imports = true

# Gradually enable stricter checking per module
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false

# ===== pytest Configuration =====
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",                    # Show summary of all test outcomes
    "-q",                     # Quiet output
    "--strict-markers",       # Ensure markers are defined
    "--strict-config",        # Ensure config is valid
    "--cov=app",              # Coverage for app directory
    "--cov=domains",          # Coverage for domains directory
    "--cov-branch",           # Branch coverage
    "--cov-report=term-missing", # Show missing lines
    "--cov-report=html",      # Generate HTML coverage report
    "--cov-report=xml",       # Generate XML for CI
]
markers = [
    "unit: Unit tests that don't require external services",
    "service: Service tests that require Neo4j or other services",
    "integration: Integration tests across multiple components",
    "slow: Tests that take a long time to run",
]
asyncio_mode = "auto"

# ===== Coverage Configuration =====
[tool.coverage.run]
source = ["app", "domains"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
    "*/conftest.py",
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abstract",
]

[tool.coverage.html]
directory = "htmlcov"

# ===== Bandit Configuration =====
[tool.bandit]
exclude_dirs = ["/tests", "/scripts"]
skips = ["B101"]  # Skip assert_used check (common in tests)
# B201-B202: Flask injection checks (not using Flask)
# B301: Pickle usage (Neo4j driver uses it safely)
# B601-B607: Shell injection checks (subprocess usage)
tests = ["B201", "B301", "B302", "B303", "B304", "B305", "B306", "B307", "B308", "B309", "B310",
         "B311", "B312", "B313", "B314", "B315", "B316", "B317", "B318", "B319", "B320", "B321",
         "B322", "B323", "B324", "B401", "B402", "B403", "B404", "B405", "B406", "B407", "B408",
         "B409", "B410", "B411", "B412", "B413", "B501", "B502", "B503", "B504", "B505", "B506",
         "B507", "B508", "B509", "B601", "B602", "B603", "B604", "B605", "B606", "B607", "B608",
         "B609", "B610", "B611", "B701", "B702", "B703"]
